backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.25.2
websockets==12.0
redis==5.0.1
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
docker==6.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
crewai
crewai-tools
pytest==7.4.3
pytest-asyncio==0.21.1
